pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.8",
    "uvicorn>=0.39.0",
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
]
